[World] Update default world values and correct gridSize type.

+++ Ibex/include/world/world.h
@@ -25,7 +25,7 @@ struct WorldOptions {
 };
 
 struct GridOptions {
-	float gridSize;
+	int gridSize;
 	glm::vec4 gridColor;
 };
